moved back some logic to starknet-types
