+++ crates/bin/starknet-sierra-upgrade-validate/src/main.rs
@@ -6,9 +6,7 @@ use anyhow::Context;
 use cairo_lang_starknet::allowed_libfuncs::{
     validate_compatible_sierra_version, AllowedLibfuncsError, ListSelector,
 };
-use cairo_lang_starknet::casm_contract_class_from_contract_class::{
-    casm_contract_class_from_contract_class, StarknetSierraCompilationError,
-};
+use cairo_lang_starknet::casm_contract_class::{CasmContractClass, StarknetSierraCompilationError};
 use cairo_lang_starknet::compiler_version::VersionId;
 use cairo_lang_starknet::contract_class::{ContractClass, ContractEntryPoints};
 use cairo_lang_utils::bigint::BigUintAsHex;
@@ -209,7 +207,7 @@ fn run_single(mut sierra_class: ContractClassInfo, config: &RunConfig) -> RunRes
         return RunResult::ValidationFailure(ValidationFailure { class_hash, err });
     };
     let compiled_contract_class =
-        match casm_contract_class_from_contract_class(contract_class, false) {
+        match CasmContractClass::from_contract_class(contract_class, false) {
             Ok(compiled_contract_class) => compiled_contract_class,
             Err(err) => {
                 return RunResult::CompilationFailure(CompilationFailure { class_hash, err });
