+++ crates/cairo-lang-starknet-types/Cargo.toml
@@ -17,20 +17,49 @@ hashbrown = { workspace = true }
 
 # Optional std
 cairo-lang-sierra = { path = "../cairo-lang-sierra", version = "2.5.3", optional = true }
+cairo-lang-sierra-to-casm = { path = "../cairo-lang-sierra-to-casm", version = "2.5.3", optional = true }
+cairo-lang-lowering = { path = "../cairo-lang-lowering", version = "2.5.3", optional = true }
+cairo-lang-semantic = { path = "../cairo-lang-semantic", version = "2.5.3", optional = true }
+cairo-lang-diagnostics = { path = "../cairo-lang-diagnostics", version = "2.5.3", optional = true }
+once_cell = { workspace = true, optional = true }
+thiserror = { workspace = true, optional = true }
+convert_case = { workspace = true, optional = true }
+num-integer = { workspace = true, optional = true }
+num-traits = { workspace = true, optional = true }
+smol_str = { workspace = true, optional = true }
+anyhow = { workspace = true, optional = true }
+sha3 = { workspace = true, optional = true }
 
 # Optional serde
 serde = { workspace = true, optional = true }
 serde_json = { workspace = true, optional = true }
 
+[dev-dependencies]
+cairo-lang-test-utils = { path = "../cairo-lang-test-utils", features = ["testing"] }
+pretty_assertions.workspace = true
+test-case.workspace = true
+
 [features]
 default = ["std"]
 std = [
-  "dep:cairo-lang-sierra",
   "cairo-felt/std",
   "cairo-lang-casm/std",
   "cairo-lang-utils/std",
   "num-bigint/std",
   "starknet-crypto/std",
+  "dep:cairo-lang-sierra",
+  "dep:cairo-lang-sierra-to-casm",
+  "dep:cairo-lang-lowering",
+  "dep:cairo-lang-semantic",
+  "dep:cairo-lang-diagnostics",
+  "dep:once_cell",
+  "dep:thiserror",
+  "dep:convert_case",
+  "dep:num-integer",
+  "dep:num-traits",
+  "dep:smol_str",
+  "dep:anyhow",
+  "dep:sha3",
 ]
 serde = ["dep:serde", "dep:serde_json", "cairo-lang-casm/serde", "cairo-lang-utils/serde"]
