+++ crates/cairo-lang-starknet-types/src/allowed_libfuncs_lists/all.json
