+++ crates/cairo-lang-starknet-types/src/allowed_libfuncs_lists/experimental.json
