+++ crates/cairo-lang-starknet-types/src/felt252_serde.rs
@@ -19,7 +19,6 @@ use cairo_lang_sierra::program::{
     Function, FunctionSignature, GenericArg, Invocation, LibfuncDeclaration, Param, Program,
     Statement, StatementIdx, TypeDeclaration,
 };
-use cairo_lang_starknet_types::compiler_version::VersionId;
 use cairo_lang_utils::bigint::BigUintAsHex;
 use cairo_lang_utils::ordered_hash_set::OrderedHashSet;
 use cairo_lang_utils::unordered_hash_map::UnorderedHashMap;
@@ -29,12 +28,9 @@ use once_cell::sync::Lazy;
 use smol_str::SmolStr;
 use thiserror::Error;
 
-use crate::contract::starknet_keccak;
+use crate::compiler_version::VersionId;
 use crate::felt252_vec_compression::{compress, decompress};
-
-#[cfg(test)]
-#[path = "felt252_serde_test.rs"]
-mod test;
+use crate::keccak::starknet_keccak;
 
 #[derive(Error, Debug, Eq, PartialEq)]
 pub enum Felt252SerdeError {
@@ -92,7 +88,7 @@ pub fn sierra_from_felt252s(
 }
 
 /// Trait for serializing and deserializing into a felt252 vector.
-trait Felt252Serde: Sized {
+pub trait Felt252Serde: Sized {
     fn serialize(&self, output: &mut Vec<BigUintAsHex>) -> Result<(), Felt252SerdeError>;
     fn deserialize(input: &[BigUintAsHex]) -> Result<(Self, &[BigUintAsHex]), Felt252SerdeError>;
 }
@@ -179,7 +175,7 @@ impl Felt252Serde for StatementIdx {
 // Impls for generic ids.
 const SHORT_STRING_BOUND: usize = 31;
 /// A set of all the supported long generic ids.
-static SERDE_SUPPORTED_LONG_IDS: Lazy<OrderedHashSet<&'static str>> = Lazy::new(|| {
+pub static SERDE_SUPPORTED_LONG_IDS: Lazy<OrderedHashSet<&'static str>> = Lazy::new(|| {
     OrderedHashSet::from_iter([
         StorageAddressFromBaseAndOffsetLibfunc::STR_ID,
         ContractAddressTryFromFelt252Libfunc::STR_ID,
