+++ crates/cairo-lang-starknet-types/src/keccak.rs
@@ -1,0 +1,13 @@
+use num_bigint::BigUint;
+use sha3::{Digest, Keccak256};
+
+/// A variant of eth-keccak that computes a value that fits in a Starknet field element.
+pub fn starknet_keccak(data: &[u8]) -> BigUint {
+    let mut hasher = Keccak256::new();
+    hasher.update(data);
+    let mut result = hasher.finalize();
+
+    // Truncate result to 250 bits.
+    *result.first_mut().unwrap() &= 3;
+    BigUint::from_bytes_be(&result)
+}
