+++ crates/cairo-lang-starknet-types/src/lib.rs
@@ -4,6 +4,15 @@
 extern crate alloc;
 
 pub mod abi;
+#[cfg(feature = "std")]
+pub mod allowed_libfuncs;
 pub mod casm_contract_class;
 pub mod compiler_version;
 pub mod contract_class;
+#[cfg(feature = "std")]
+pub mod keccak;
+
+#[cfg(feature = "std")]
+pub mod felt252_serde;
+#[cfg(feature = "std")]
+mod felt252_vec_compression;
