+++ crates/cairo-lang-starknet/src/plugin/storage.rs
@@ -1,5 +1,6 @@
 use cairo_lang_defs::patcher::RewriteNode;
 use cairo_lang_defs::plugin::PluginDiagnostic;
+use cairo_lang_starknet_types::keccak::starknet_keccak;
 use cairo_lang_syntax::node::db::SyntaxGroup;
 use cairo_lang_syntax::node::{ast, Terminal, TypedSyntaxNode};
 use cairo_lang_utils::try_extract_matches;
@@ -12,7 +13,6 @@ use super::consts::{
 use super::starknet_module::generation_data::StarknetModuleCommonGenerationData;
 use super::starknet_module::StarknetModuleKind;
 use super::utils::has_v0_attribute;
-use crate::contract::starknet_keccak;
 
 /// Generate getters and setters for the members of the storage struct.
 pub fn handle_storage_struct(
